feat(structure): control column visibility in sheet list
